refactor(renderer): extract launcher keyboard controls
